Use new GraphQL API for V2 site
