+++ client/relay.config.js
@@ -1,6 +1,6 @@
 module.exports = {
   src: "./src",
-  schema: "../server/schema.graphql",
+  schema: "src/queries/schema.graphql",
   exclude: ["**/node_modules/**", "**/__mocks__/**", "**/__generated__/**"],
   language: "typescript",
   artifactDirectory: "src/queries/__generated__",
