+++ client/src/lib/client/relay_client_environment.ts
@@ -3,12 +3,12 @@ import { Environment, Network, RecordSource, Store } from "relay-runtime";
 const baseUrl = new URL(
   process.env.NODE_ENV === "production"
     ? "https://edhtop16.com"
-    : "http://localhost:8000",
+    : "http://localhost:3000",
 );
 
 export function createClientNetwork() {
   return Network.create(async (params, variables) => {
-    const response = await fetch(new URL("/api/v2/graphql", baseUrl), {
+    const response = await fetch(new URL("/v2/api/graphql", baseUrl), {
       method: "POST",
       credentials: "include",
       headers: {
